requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==4.0.3
bidict==0.23.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
fastapi==0.115.11
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.85.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-core==0.3.47
langchain-google-vertexai==2.0.16
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
msgpack==1.1.0
numpy==2.0.2
orjson==3.10.15
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shapely==2.0.7
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tenacity==9.0.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wsproto==1.2.0
zstandard==0.23.0
